i am very lost

+++ TheRegularEngineV2/Compilers/TextureCompiler/src/TextureCompiler.cpp
@@ -115,7 +115,7 @@ namespace TRE
 		//crn_free_block(outputData);
 		m_Texture = std::make_unique<Texture>();
 		m_Texture->Data = reinterpret_cast<void*>(pixels);
-		int len = std::min(descriptor.GetTextureName().length(), sizeof(m_Texture->Name));
+		int len = static_cast<int>(std::min(descriptor.GetTextureName().length(), sizeof(m_Texture->Name)));
 		for (int i = 0; i < len; ++i)
 			m_Texture->Name[i] = descriptor.GetTextureName()[i];
 		m_Texture->Width = texWidth;
